+++ include/tfblib/tfb_inline_funcs.h
@@ -6,19 +6,19 @@
 /* Essential variables */
 extern void *__fb_buffer;
 extern void *__fb_real_buffer;
-extern u32 __fb_screen_w;
-extern u32 __fb_screen_h;
+extern int __fb_screen_w;
+extern int __fb_screen_h;
 extern size_t __fb_size;
 extern size_t __fb_pitch;
 extern size_t __fb_pitch_div4; /* see the comment in drawing.c */
 
 /* Window-related variables */
-extern u32 __fb_win_w;
-extern u32 __fb_win_h;
-extern u32 __fb_off_x;
-extern u32 __fb_off_y;
-extern u32 __fb_win_end_x;
-extern u32 __fb_win_end_y;
+extern int __fb_win_w;
+extern int __fb_win_h;
+extern int __fb_off_x;
+extern int __fb_off_y;
+extern int __fb_win_end_x;
+extern int __fb_win_end_y;
 
 /* Color-related variables */
 extern u32 __fb_r_mask;
@@ -38,12 +38,12 @@ inline u32 tfb_make_color(u8 r, u8 g, u8 b)
           ((b << __fb_b_pos) & __fb_b_mask);
 }
 
-inline void tfb_draw_pixel(u32 x, u32 y, u32 color)
+inline void tfb_draw_pixel(int x, int y, u32 color)
 {
    x += __fb_off_x;
    y += __fb_off_y;
 
-   if (x < __fb_win_end_x && y < __fb_win_end_y)
+   if ((u32)x < (u32)__fb_win_end_x && (u32)y < (u32)__fb_win_end_y)
       ((volatile u32 *)__fb_buffer)[x + y * __fb_pitch_div4] = color;
 }
